requirements.txt
aiogram==3.*
asyncpg
python-dotenv